added a claim subcommand to the daily command
